+++ app/consumer.py
@@ -17,9 +17,23 @@ from structlog import wrap_logger
 from app.async_consumer import AsyncConsumer
 from app.response_processor import ResponseProcessor
 from app import settings
+from app.queue_publisher import QueuePublisher
 
 logging.basicConfig(level=settings.LOGGING_LEVEL, format=settings.LOGGING_FORMAT)
 logger = wrap_logger(logging.getLogger(__name__))
+
+from app.helpers.exceptions import BadMessageError, DecryptError, RetryableError
+
+
+def get_delivery_count_from_properties(properties):
+    """
+    Returns the delivery count for a message from the rabbit queue. The
+    value is auto-set by rabbitmq.
+    """
+    delivery_count = 0
+    if properties.headers and 'x-delivery-count' in properties.headers:
+        delivery_count = properties.headers['x-delivery-count']
+    return delivery_count + 1
 
 
 class Consumer(AsyncConsumer):
@@ -27,27 +41,50 @@ class Consumer(AsyncConsumer):
     def __init__(self, args=None, cfg=None):
         self._args = args
         self._cfg = cfg
+        self.quarantine_publisher = QueuePublisher(logger, settings.RABBIT_URLS, settings.RABBIT_QUARANTINE_QUEUE)
         super().__init__()
 
     def on_message(self, unused_channel, basic_deliver, properties, body):
-        logger.info('Received message', queue=self.QUEUE, delivery_tag=basic_deliver.delivery_tag, app_id=properties.app_id)
 
-        options = ResponseProcessor.options()
+        delivery_count = get_delivery_count_from_properties(properties)
+
+        logger.info(
+            'Received message',
+            queue=self.QUEUE,
+            delivery_tag=basic_deliver.delivery_tag,
+            delivery_count=delivery_count,
+            app_id=properties.app_id
+        )
+
+        # options = ResponseProcessor.options()
         processor = ResponseProcessor(logger)
 
         try:
-            message = body.decode("utf-8")
-            processed_ok = processor.process(message, **options)
+            processor.process(body.decode("utf-8"))
+            self.acknowledge_message(basic_deliver.delivery_tag, tx_id=processor.tx_id)
 
-            if processed_ok:
-                self.acknowledge_message(basic_deliver.delivery_tag, tx_id=processor.tx_id)
+        except DecryptError as e:
+            # Throw it into the quarantine queue to be dealt with
+            self.quarantine_publisher.publish_message(body)
+            self.reject_message(basic_deliver.delivery_tag, tx_id=processor.tx_id)
+            logger.error("Bad decrypt", action="quarantined", exception=e, tx_id=processor.tx_id, delivery_count=delivery_count)
 
-        except Exception as e:
-            logger.error("ResponseProcessor failed", exception=e, tx_id=processor.tx_id)
+        except BadMessageError as e:
+            # If it's a bad message then we have to reject it
+            self.reject_message(basic_deliver.delivery_tag, tx_id=processor.tx_id)
+            logger.error("Bad message", action="rejected", exception=e, tx_id=processor.tx_id, delivery_count=delivery_count)
+
+        except (RetryableError, Exception) as e:
+            self.nack_message(basic_deliver.delivery_tag, tx_id=processor.tx_id)
+            logger.error("Failed to process", action="nack", exception=e, tx_id=processor.tx_id, delivery_count=delivery_count)
 
 
 def main(args=None):
     logger.debug("Starting consumer")
+
+    if settings.SDX_COLLECT_SECRET is None:
+        logger.error("No SDX_COLLECT_SECRET env var supplied")
+        sys.exit(1)
 
     consumer = Consumer(args)
     try: