+++ app/response_processor.py
@@ -6,6 +6,7 @@ from requests.packages.urllib3.exceptions import MaxRetryError
 from app import settings
 from app.private_publisher import PrivatePublisher
 from app.settings import session
+from app.helpers.exceptions import DecryptError, BadMessageError, RetryableError
 
 
 class ResponseProcessor:
@@ -31,11 +32,9 @@ class ResponseProcessor:
             logger, settings.RABBIT_URLS, settings.RABBIT_CTP_RECEIPT_QUEUE
         )
 
-    def process(self, encrypted_survey, **kwargs):
+    def process(self, encrypted_survey):
         # decrypt
-        decrypt_ok, decrypted_json = self.decrypt_survey(encrypted_survey)
-        if not decrypt_ok:
-            return False
+        decrypted_json = self.decrypt_survey(encrypted_survey)
 
         metadata = decrypted_json['metadata']
         self.logger = self.logger.bind(user_id=metadata['user_id'], ru_ref=metadata['ru_ref'])
@@ -44,16 +43,12 @@ class ResponseProcessor:
             self.tx_id = decrypted_json['tx_id']
             self.logger = self.logger.bind(tx_id=self.tx_id)
 
-        # validate
-        validate_ok = self.validate_survey(decrypted_json)
-        if not validate_ok:
-            decrypted_json['invalid'] = True
+        self.validate_survey(decrypted_json)
+        self.store_survey(decrypted_json)
+        self.send_receipt(decrypted_json)
+        return
 
-        # store
-        store_ok = self.store_survey(decrypted_json)
-        if not store_ok:
-            return False
-
+    def send_receipt(self, decrypted_json):
         receipt_json = {
             'tx_id': decrypted_json['tx_id'],
             'collection': {
@@ -66,30 +61,30 @@ class ResponseProcessor:
         }
 
         if decrypted_json.get("survey_id") and decrypted_json["survey_id"] == "census":
-            queue_ok = self.ctp_publisher.publish_message(dumps(receipt_json), **kwargs)
+            queue_ok = self.ctp_publisher.publish_message(dumps(receipt_json), secret=settings.SDX_COLLECT_SECRET)
         else:
-            queue_ok = self.rrm_publisher.publish_message(dumps(receipt_json), **kwargs)
+            queue_ok = self.rrm_publisher.publish_message(dumps(receipt_json), secret=settings.SDX_COLLECT_SECRET)
 
         if not queue_ok:
-            return False
-        else:
-            return True
+            raise RetryableError()
 
     def decrypt_survey(self, encrypted_survey):
+        self.logger.debug("Decrypting survey")
         response = self.remote_call(settings.SDX_DECRYPT_URL, data=encrypted_survey)
-        decrypt_ok = self.response_ok(response)
-        if decrypt_ok:
-            return (True, response.json())
-        else:
-            return (False, None)
+        try:
+            self.response_ok(response)
+        except BadMessageError:
+            # Translate the Bad Message into a Decrypt Error to force quarantine
+            raise DecryptError
+        return response.json()
 
     def validate_survey(self, decrypted_json):
-        response = self.remote_call(settings.SDX_VALIDATE_URL, json=decrypted_json)
-        return self.response_ok(response)
+        self.logger.debug("Validating survey")
+        self.response_ok(self.remote_call(settings.SDX_VALIDATE_URL, json=decrypted_json))
 
     def store_survey(self, decrypted_json):
-        response = self.remote_call(settings.SDX_STORE_URL, json=decrypted_json)
-        return self.response_ok(response)
+        self.logger.debug("Storing survey")
+        self.response_ok(self.remote_call(settings.SDX_STORE_URL, json=decrypted_json))
 
     def remote_call(self, request_url, json=None, data=None, headers=None, verify=True, auth=None):
         try:
@@ -109,10 +104,17 @@ class ResponseProcessor:
             self.logger.error("Max retries exceeded (5)", request_url=request_url)
 
     def response_ok(self, res):
+        res_logger = self.logger
+        res_logger.bind(request_url=res.url, status_code=res.status_code)
+
         if res.status_code == 200 or res.status_code == 201:
-            self.logger.info("Returned from service", request_url=res.url, status_code=res.status_code)
-            return True
+            res_logger.info("Returned from service", response="ok")
+            return
+
+        elif res.status_code == 400:
+            res_logger.info("Returned from service", response="client error")
+            raise BadMessageError
 
         else:
-            self.logger.error("Returned from service", request_url=res.url, status_code=res.status_code)
-            return False
+            res_logger.error("Returned from service", response="service error")
+            raise RetryableError