+++ tests/test_response_processor.py
@@ -2,15 +2,28 @@ import json
 import logging
 import os
 import unittest
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, Mock
+import mock
+from requests import Response
 
 from structlog import wrap_logger
 
 from app.response_processor import ResponseProcessor
-from tests.test_data import fake_encrypted, valid_decrypted
+from tests.test_data import valid_decrypted
+from app.helpers.exceptions import DecryptError, RetryableError, BadMessageError
+from app import settings
 
 logger = wrap_logger(logging.getLogger(__name__))
 valid_json = json.loads(valid_decrypted)
+
+
+class RRMQueue(Exception):
+    # Hacky, but using exception to check something is on the RIGHT queue in the test
+    pass
+
+
+class CTPQueue(Exception):
+    pass
 
 
 class TestResponseProcessorSettings(unittest.TestCase):
@@ -35,35 +48,119 @@ class TestResponseProcessorSettings(unittest.TestCase):
 
 class TestResponseProcessor(unittest.TestCase):
 
-    def test_decrypt_failure(self):
-        rp = ResponseProcessor(logger)
-        rp.decrypt_survey = MagicMock(return_value=(False, None))
-        response = rp.process(fake_encrypted)
+    def setUp(self):
+        self.rp = ResponseProcessor(logger)
 
-        rp.decrypt_survey.assert_called_with(fake_encrypted)
-        self.assertFalse(response)
+    def _process(self):
+        self.rp.process("NxjsJBSahBXHSbxHBasx")
 
-    def test_decrypt_success_validate_failure(self):
-        rp = ResponseProcessor(logger)
-        rp.decrypt_survey = MagicMock(return_value=(True, valid_json))
-        rp.validate_survey = MagicMock(return_value=False)
-        rp.store_survey = MagicMock(return_value=True)
-        rp.rrm_publisher.publish_message = MagicMock(return_value=True)
-        rp.skip_receipt = True
-        response = rp.process(fake_encrypted)
+    def test_decrypt(self):
+        # <decrypt>
+        self.rp.validate_survey = MagicMock()
+        self.rp.store_survey = MagicMock()
+        self.rp.send_receipt = MagicMock()
 
-        rp.decrypt_survey.assert_called_with(fake_encrypted)
-        # When validate returns a failure, the process still actually is meant
-        # to continue, store and return successfully. So even though it's a
-        # "failure" we still expect True
-        self.assertTrue(response)
+        r = Response()
 
-    def test_validate_success_store_failure(self):
-        rp = ResponseProcessor(logger)
-        rp.decrypt_survey = MagicMock(return_value=(True, valid_json))
-        rp.validate_survey = MagicMock(return_value=True)
-        rp.store_survey = MagicMock(return_value=False)
-        response = rp.process(fake_encrypted)
+        with mock.patch('app.response_processor.ResponseProcessor.remote_call') as call_mock:
+            call_mock.return_value = r
+            r._content = valid_decrypted.encode('utf-8')
 
-        rp.store_survey.assert_called_with(valid_json)
-        self.assertFalse(response)
+            # 500 - retry
+            r.status_code = 500
+            with self.assertRaises(RetryableError):
+                self._process()
+
+            # 400 - bad
+            r.status_code = 400
+            with self.assertRaises(DecryptError):
+                self._process()
+
+            # 200 - ok
+            r.status_code = 200
+            self._process()
+
+    def test_validate(self):
+        self.rp.decrypt_survey = MagicMock(return_value=valid_json)
+        # <validate>
+        self.rp.store_survey = MagicMock()
+        self.rp.send_receipt = MagicMock()
+
+        r = Response()
+        with mock.patch('app.response_processor.ResponseProcessor.remote_call') as call_mock:
+            call_mock.return_value = r
+
+            # 500 - retry
+            r.status_code = 500
+            with self.assertRaises(RetryableError):
+                self._process()
+
+            # 400 - bad
+            r.status_code = 400
+            with self.assertRaises(BadMessageError):
+                self._process()
+
+            # 200 - ok
+            r.status_code = 200
+            self._process()
+
+    def test_store(self):
+        self.rp.decrypt_survey = MagicMock(return_value=valid_json)
+        self.rp.validate_survey = MagicMock()
+        # <store>
+        self.rp.send_receipt = MagicMock()
+
+        r = Response()
+        with mock.patch('app.response_processor.ResponseProcessor.remote_call') as call_mock:
+            call_mock.return_value = r
+
+            # 500 - retry
+            r.status_code = 500
+            with self.assertRaises(RetryableError):
+                self._process()
+
+            # 400 - bad
+            r.status_code = 400
+            with self.assertRaises(BadMessageError):
+                self._process()
+
+            # 200 - ok
+            r.status_code = 200
+            self._process()
+
+    def test_send_receipt(self):
+        self.rp.decrypt_survey = MagicMock(return_value=valid_json)
+        self.rp.validate_survey = MagicMock()
+        self.rp.store_survey = MagicMock()
+        # <send_receipt>
+
+        # Bad key - none set (shouldn't occur as service will not start without key)
+        settings.SDX_COLLECT_SECRET = None
+        with self.assertRaises(TypeError):
+            self._process()
+
+        # Subsequent tests expect valid key
+        settings.SDX_COLLECT_SECRET = "seB388LNHgxcuvAcg1pOV20_VR7uJWNGAznE0fOqKxg=".encode('ascii')
+
+        # rrm queue fail
+        with self.assertRaises(RetryableError):
+            self._process()
+
+        # rrm publish ok
+        self.rp.rrm_publisher.publish_message = MagicMock()
+        self._process()
+
+        self._process()
+
+        # Queue types
+        self.rp.rrm_publisher.publish_message = Mock(side_effect=RRMQueue)
+        self.rp.ctp_publisher.publish_message = Mock(side_effect=CTPQueue)
+
+        with self.assertRaises(RRMQueue):
+            self.rp.send_receipt(valid_json)
+
+        census_json = valid_json
+        census_json['survey_id'] = 'census'
+
+        with self.assertRaises(CTPQueue):
+            self.rp.send_receipt(valid_json)